+++ backend/canisters/storage_bucket/impl/src/model/files.rs
@@ -5,6 +5,7 @@ use crate::model::stable_blob_storage::StableBlobStorage;
 use crate::{calc_chunk_count, MAX_BLOB_SIZE_BYTES};
 use candid::Principal;
 use serde::{Deserialize, Serialize};
+use stable_memory_map::StableMemoryMap;
 use std::cmp::Ordering;
 use std::collections::btree_map::Entry::{Occupied, Vacant};
 use std::collections::{BTreeMap, BTreeSet, VecDeque};
@@ -207,7 +208,8 @@ impl Files {
             mime_type: file.mime_type,
         };
 
-        self.files.set(new_file_id, new_file);
+        self.files.insert(new_file_id, new_file);
+
         ForwardFileResult::Success(FileAdded {
             file_id: new_file_id,
             hash,
@@ -237,7 +239,7 @@ impl Files {
                         files_removed.push(file_removed);
                     }
                 } else {
-                    self.files.set(file_id, file);
+                    self.files.insert(file_id, file);
                 }
             }
 
@@ -252,7 +254,7 @@ impl Files {
     pub fn update_owner(&mut self, file_id: &FileId, new_owner: Principal) -> bool {
         if let Some(mut file) = self.get(file_id) {
             file.owner = new_owner;
-            self.files.set(*file_id, file);
+            self.files.insert(*file_id, file);
             true
         } else {
             false
@@ -265,7 +267,7 @@ impl Files {
             if let Some(mut file) = self.get(file_id) {
                 if file.accessors.remove(&old_accessor_id) {
                     file.accessors.insert(new_accessor_id);
-                    self.files.set(*file_id, file);
+                    self.files.insert(*file_id, file);
                     self.accessors_map.link(new_accessor_id, *file_id);
                 }
             }
@@ -332,7 +334,7 @@ impl Files {
             self.expiration_queue.entry(expiry).or_default().push_back(file_id);
         }
 
-        self.files.set(
+        self.files.insert(
             file_id,
             File {
                 owner: completed_file.owner,
@@ -345,7 +347,7 @@ impl Files {
     }
 
     fn remove_file(&mut self, file_id: FileId) -> Option<FileRemoved> {
-        let file = self.files.remove(&file_id)?;
+        let file = self.files.remove(&file_id)?.into_value();
 
         if self.reference_counts.decr(file.hash) == 0 {
             self.remove_blob(&file.hash);